pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mini-fred"
version = "0.1.0"
description = "Minimal, reproducible Mini-FRED warehouse scaffolding"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "duckdb>=0.10",
    "pandas>=2.2",
    "pyyaml>=6.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0"
]
